+++ ntex/src/http/h1/dispatcher.rs
@@ -838,7 +838,7 @@ where
 #[cfg(test)]
 mod tests {
     use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-    use std::{cell::Cell, future::poll_fn, io, sync::Arc};
+    use std::{cell::Cell, future::poll_fn, future::Future, io, sync::Arc};
 
     use ntex_h2::Config;
     use rand::Rng;
@@ -848,7 +848,7 @@ mod tests {
     use crate::http::h1::{ClientCodec, DefaultControlService};
     use crate::http::{body, Request, ResponseHead, StatusCode};
     use crate::io::{self as nio, Base};
-    use crate::service::{boxed, fn_service, IntoService};
+    use crate::service::{fn_service, IntoService};
     use crate::util::{lazy, stream_recv, Bytes, BytesMut};
     use crate::{codec::Decoder, testing::Io, time::sleep, time::Millis, time::Seconds};