+++ ntex/src/web/service.rs
@@ -389,7 +389,6 @@ tuple_web_service!((0,A),(1,B),(2,C),(3,D),(4,E),(5,F),(6,G),(7,H),(8,I),(9,J),(
 mod tests {
     use super::*;
     use crate::http::{Method, StatusCode};
-    use crate::service::fn_service;
     use crate::web::test::{init_service, TestRequest};
     use crate::web::{self, guard, App, DefaultError, HttpResponse};