+++ ntex/tests/http_awc_rustls_client.rs
@@ -62,7 +62,7 @@ async fn test_connection_reuse_h2() {
 
     let srv = test_server(move || {
         let num2 = num2.clone();
-        chain_factory(dev::map_app_err(move |io| {
+        chain_factory(dev::map_box_err(move |io| {
             num2.fetch_add(1, Ordering::Relaxed);
             Ready::Ok(io)
         }))