+++ ntex/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [1.1.0] - 2024-02-06
+
+* http: Add http/1 control service
+
+* http: Expose http/2 control service
+
 ## [1.0.0] - 2024-01-09
 
 * web: Use async fn for Responder and Handler traits